pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "data-analysis-agent"
version = "2.0.0"
description = "基于 LangGraph 的智能数据分析 Agent"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Data Agent Team"}
]
keywords = ["langgraph", "agent", "data-analysis", "mysql", "llm"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "langgraph>=0.2.0",
    "langchain-openai>=0.1.0",
    "langchain-core>=0.2.0",
    "pymysql>=1.1.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "flake8>=7.0.0",
    "mypy>=1.8.0",
]

[project.urls]
Homepage = "https://github.com/example/data-analysis-agent"
Repository = "https://github.com/example/data-analysis-agent"

[project.scripts]
agent-api = "scripts.run_api:main"
agent-smoke = "scripts.run_smoke:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["config*", "core*", "database*", "models*", "tools*", "agents*", "api*", "services*", "middleware*", "memory*", "filesystem*", "scripts*", "utils*"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
